Core geometry methods now in geo module
Swarm has been getting huge and the number of static methods was blowing
up. It makes more sense to organize these key geometry functions in a
separate module. Additionally, since these functions are the key
workhorse of Planktos, it will make them easier to rewrite in C or
Cython down the road, or to throw on a GPU if that is desirable
at some point.
