+++ tests/visualtest_2d.py
@@ -57,10 +57,10 @@ print('Moving swarm...')
 for ii in range(180): # 500
     s.move(0.0005)
 
-s.plot()
-for ii in range(5): # 500
-    s.move(0.0005)
-    s.plot()
+# s.plot()
+# for ii in range(5): # 500
+#     s.move(0.0005)
+#     s.plot()
 
 # s.plot()
 # s.plot_all(movie_filename='leaf_2d_vort.mp4', figsize=(6,9), fps=30, fluid='vort')
